decouple ImportsKeyYamlToPhpFactory
